ADD Search at all in CSW server (search text empty) REF Button search enabled after check a CSW server REF Page size for summary records search to 10 REF Minor

+++ geoplatform-gui/core/geoplatform-widget/finder-widget/src/main/java/org/geosdi/geoplatform/gui/server/service/impl/GPCatalogFinderService.java
@@ -148,7 +148,7 @@ public class GPCatalogFinderService implements IGPCatalogFinderService {
             HttpServletRequest httpServletRequest)
             throws GeoPlatformException {
 
-        Long serversCount;
+        int serversCount;
         ArrayList<SummaryRecord> searchRecords;
         try {
             serversCount = geoPlatformCSWClient.getSummaryRecordsCount(catalogFinder);
@@ -174,7 +174,7 @@ public class GPCatalogFinderService implements IGPCatalogFinderService {
         }
 
         return new BasePagingLoadResult<SummaryRecord>(searchRecords,
-                config.getOffset(), serversCount.intValue());
+                config.getOffset(), serversCount);
     }
 
     /**
